add: operator +=, -= for MyCollection and == for YoutubeChannel
When defining -= operator for MyCollection we use".remove(YoutubeChannel ytChannel)" method. This method deletes all the elements that satisfy the equal operator with the given argument in the entire list.

Also the function implementing the equal operator between YoutubeChannel types must be a constant argument and a constant function. (not sure what it means)